requirements.txt
alpaca-py>=0.9.0
yfinance>=0.2.40
streamlit>=1.40.0
ta>=0.10.2
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
